+++ frontend/src/features/discussion/ui/DiscussionCard/DiscussionSmall.tsx
@@ -5,10 +5,14 @@ import { vars } from '@/theme/index.css';
 import { formatDateToTimeString } from '@/utils/date/format';
 
 import type { DiscussionDTO } from '../../model';
+import DiscussionConfirmButton from '../DiscussionConfirmButton';
 import { containerStyle } from './card.css';
 
+// 추후 삭제할 코드가 포함되어 있습니다.
+// TODO: 일정 확정 버튼 디테일 페이지로 이동
+// eslint-disable-next-line
 export const DiscussionSmall = (
-  { discussion, onClick }: { discussion: DiscussionDTO; onClick: () => void },
+  { discussion }: { discussion: DiscussionDTO },
 ) => {
   const ADJUSTMENT_LENGTH = discussion.usersForAdjust.length;
   const isRecommend = ADJUSTMENT_LENGTH === 0;
@@ -17,9 +21,12 @@ export const DiscussionSmall = (
       className={containerStyle({ isRecommend })}
       direction='column'
       gap={300}
-      onClick={onClick}
       width='100%'
     >
+      <DiscussionConfirmButton 
+        endDateTime={discussion.endDateTime}
+        startDateTime={discussion.startDateTime}
+      />
       {isRecommend && <Chip color='blue' size='lg'>추천</Chip>}
       <Flex direction='column' gap={50}>
         <Text color={vars.color.Ref.Netural[600]} typo='b3R'>
