+++ frontend/src/features/discussion/ui/DiscussionCard/card.css.ts
@@ -12,8 +12,6 @@ export const containerStyle = recipe({
     borderLeft: '3px solid transparent',
 
     backgroundColor: vars.color.Ref.Netural.White,
-
-    cursor: 'pointer',
   },
   variants: {
     isRecommend: {
@@ -36,8 +34,6 @@ export const largeContainerStyle = style({
   borderRadius: vars.radius[500],
 
   backgroundColor: vars.color.Ref.Netural.White,
-
-  cursor: 'pointer',
 });
 
 export const rankContainerStyle = recipe({
