+++ frontend/src/features/discussion/ui/DiscussionCard/index.tsx
@@ -1,6 +1,3 @@
-import { useParams } from '@tanstack/react-router';
-
-import { useDiscussionConfirmMutation } from '../../api/mutations';
 import type { DiscussionDTO } from '../../model';
 import { DiscussionLarge } from './DiscussionLarge';
 import { DiscussionSmall } from './DiscussionSmall';
@@ -11,33 +8,15 @@ interface DiscussionCardProps {
   rank?: number;
 }
 
-const DiscussionCard = ({ size, discussion, rank }: DiscussionCardProps) => {
-  const param: { id: string } = useParams({ from: '/_main/discussion/$id' });
-  const { mutate } = useDiscussionConfirmMutation();
-
-  const handleClickConfirm = () => {
-    if (!param.id) return;
-    mutate({
-      id: param.id, 
-      body: {
-        startDateTime: discussion.startDateTime,
-        endDateTime: discussion.endDateTime,
-      }, 
-    });
-  };
-
-  return (
-    size === 'lg' ? 
-      <DiscussionLarge
-        discussion={discussion}
-        onClick={handleClickConfirm}
-        rank={rank as number}
-      /> : 
-      <DiscussionSmall 
-        discussion={discussion}
-        onClick={handleClickConfirm}
-      />
-  );
-};
+const DiscussionCard = ({ size, discussion, rank }: DiscussionCardProps) => (
+  size === 'lg' ? 
+    <DiscussionLarge
+      discussion={discussion}
+      rank={rank as number}
+    /> : 
+    <DiscussionSmall 
+      discussion={discussion}
+    />
+);
 
 export default DiscussionCard;
