+++ frontend/src/features/discussion/ui/DiscussionConfirmButton/index.tsx
@@ -1,0 +1,27 @@
+import { useParams } from '@tanstack/react-router';
+
+import Button from '@/components/Button';
+
+import { useDiscussionConfirmMutation } from '../../api/mutations';
+import type { DiscussionDTO } from '../../model';
+
+const DiscussionConfirmButton = (
+  { startDateTime, endDateTime }: Omit<DiscussionDTO, 'usersForAdjust'>,
+) => {
+  const param: { id: string } = useParams({ from: '/_main/discussion/$id' });
+  const { mutate } = useDiscussionConfirmMutation();
+  
+  const handleClickConfirm = () => {
+    if (!param.id) return;
+    mutate({
+      id: param.id, 
+      body: { startDateTime, endDateTime }, 
+    });
+  };
+
+  return (
+    <Button onClick={handleClickConfirm}>일정 확정하기</Button>
+  );
+};
+
+export default DiscussionConfirmButton;
